update npm and bower dependencies/update to latest ember-cli 2.2.0-beta.1 and run ember init/transition Brocfile.js to ember-cli-build.js
