+++ bower.json
@@ -3,22 +3,20 @@
   "version": "VERSION_STRING_PLACEHOLDER",
   "private": true,
   "dependencies": {
-    "ember": "^2.0.0",
-    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
-    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
-    "ember-data": "~2.1.0",
-    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-qunit": "0.3.3",
-    "ember-qunit-notifications": "0.0.7",
-    "ember-resolver": "~0.1.15",
+    "ember": "^2.2.0",
+    "ember-cli-shims": "0.0.6",
+    "ember-cli-test-loader": "0.2.1",
+    "ember-data": "~2.2.1",
+    "ember-load-initializers": "0.1.7",
+    "ember-qunit-notifications": "0.1.0",
     "jquery": "^1.11.1",
-    "qunit": "~1.17.1",
-    "pretender": "^0.9.0"
+    "qunit": "~1.20.0",
+    "pretender": "^0.10.1"
   },
   "devDependencies": {
-    "loader.js": "~3.2.1"
+    "loader.js": "~3.5.0"
   },
   "resolutions": {
-    "ember": "^2.0.0"
+    "ember": "^2.2.0"
   }
 }
