+++ package.json
@@ -18,29 +18,33 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "broccoli-asset-rev": "^2.0.2",
-    "broccoli-babel-transpiler": "^5.1.1",
+    "broccoli-asset-rev": "^2.4.1",
+    "broccoli-babel-transpiler": "^5.5.0",
     "broccoli-concat": "0.0.13",
     "broccoli-es3-safe-recast": "^2.0.0",
-    "broccoli-funnel": "^0.2.3",
-    "broccoli-merge-trees": "^0.2.1",
-    "broccoli-replace": "^0.3.1",
-    "broccoli-stew": "^0.3.2",
-    "ember-cli": "^1.13.0",
-    "ember-cli-app-version": "0.3.3",
+    "broccoli-funnel": "^1.0.1",
+    "broccoli-merge-trees": "^1.1.0",
+    "broccoli-replace": "^0.11.0",
+    "broccoli-stew": "^1.0.4",
+    "ember-cli": "^2.2.0-beta.1",
+    "ember-cli-app-version": "^1.0.0",
     "ember-cli-content-security-policy": "0.4.0",
-    "ember-cli-dependency-checker": "^1.0.0",
-    "ember-cli-htmlbars": "0.7.6",
-    "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-pretender": "0.4.0",
-    "ember-cli-qunit": "0.3.13",
-    "ember-cli-uglify": "^1.0.1",
-    "ember-data": "^2.0.0",
-    "ember-disable-prototype-extensions": "^1.0.0",
-    "ember-disable-proxy-controllers": "^1.0.0",
-    "ember-export-application-global": "^1.0.2",
-    "ember-try": "0.0.7",
-    "ember-watson": "^0.5.7",
+    "ember-cli-dependency-checker": "^1.2.0",
+    "ember-cli-htmlbars": "^1.0.1",
+    "ember-cli-htmlbars-inline-precompile": "^0.3.1",
+    "ember-cli-inject-live-reload": "^1.4.0",
+    "ember-cli-pretender": "0.5.0",
+    "ember-cli-qunit": "^1.1.0",
+    "ember-cli-release": "0.2.8",
+    "ember-cli-sri": "^1.2.0",
+    "ember-cli-uglify": "^1.2.0",
+    "ember-data": "^2.2.1",
+    "ember-disable-prototype-extensions": "^1.0.1",
+    "ember-disable-proxy-controllers": "^1.0.1",
+    "ember-export-application-global": "^1.0.5",
+    "ember-resolver": "^2.0.3",
+    "ember-try": "~0.0.8",
+    "ember-watson": "^0.7.0",
     "github": "^0.2.4",
     "rsvp": "^3.1.0"
   },
@@ -48,7 +52,7 @@
     "ember-addon"
   ],
   "dependencies": {
-    "ember-cli-babel": "^5.0.0"
+    "ember-cli-babel": "^5.1.6"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
